clean setup.py dependencies
For windows build, the order of dependencies appears to affect pip installation.
There is a (minor) conflict for pyscard dependency, as pysatochip requires v>=1.9.9 (resolving to latest 2.0.7), while openpgp requires exactly 2.0.3.
If v2.0.7 is installed before 2.03, it generates an error and some unrelated dependencies (h11, cffi,...) are not installed for an unknown reason.
By simply moving pysatochip up in 'install_requires' list, pyscard v2.0.3 is installed and conflict is avoided...